Renames ErrorUtil to EspressoTestUtil (and includes setChecked())
