+++ app/src/androidTest/java/io/github/alansanchezp/gnomy/account/ArchivedAccountsDialogFragmentInstrumentedTest.java
@@ -24,7 +24,7 @@ import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withEffectiveVisibility;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
-import static io.github.alansanchezp.gnomy.ErrorUtil.assertThrows;
+import static io.github.alansanchezp.gnomy.EspressoTestUtil.assertThrows;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
